+++ src/main/java/racingcar/domain/car/dto/output/CarsFinishedDto.java
@@ -1,0 +1,8 @@
+package racingcar.domain.car.dto.output;
+
+import java.util.List;
+
+public record CarsFinishedDto(
+        List<String> winnerNames
+) {
+}
