--- src/main/java/racingcar/domain/game/MoveResult.java
@@ -1,7 +1,0 @@
-package racingcar.domain.game;
-
-public record MoveResult(
-        String carName,
-        int coordinate
-) {
-}
