+++ src/test/java/racingcar/domain/move/CoordinateTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.coordinate.Coordinate;
 import racingcar.exception.InvalidCoordinateException;
 
 final class CoordinateTest {
